feat(frontend): add database tab to workspace settings dialog
