+++ frontend/src/app/(console)/layout.tsx
@@ -49,12 +49,13 @@ export default function ConsoleLayout({
               >
                 {t('llmSettingsNav')}
               </button>
-              <Link
-                href="/console/database"
+              <button
+                type="button"
                 className="rounded-full border border-cyan-300/35 bg-cyan-300/10 px-4 py-2 text-sm text-cyan-800 transition hover:bg-cyan-300/20 dark:text-cyan-100"
+                onClick={() => openWorkspaceSettings('database')}
               >
                 {t('databaseAdminNav')}
-              </Link>
+              </button>
             </div>
             <Link
               href="/console"
